[Fix]loadHistoryData버튼을 loadSportsData로 재활용 가능하게 변경
