+++ src/main/java/com/multicampus/matchcode/model/entity/MemberAndPointDTO.java
@@ -8,6 +8,7 @@ public class MemberAndPointDTO { //mypage 메인화면에 이름과 포인트를
     private MemberDTO memberDTO; //세션에서 꺼내서 쓰면 됨
     private PointDTO pointDTO;
 
+
     public MemberAndPointDTO(MemberDTO memberDTO, PointDTO pointDTO) {
         this.memberDTO = memberDTO;
         this.pointDTO = pointDTO;
