+++ src/main/resources/templates/khj/matchhistory.html
@@ -6,14 +6,14 @@
 </head>
 <body>
 <script src="https://code.jquery.com/jquery-3.4.1.js"></script>
-<script src="js/historybutton.js"></script>
+<script src="js/sportsbutton.js"></script>
 <h2 style="display:inline">매치 히스토리</h2>
 <button id="futsal">풋살</button>
 <button id="basketball">농구</button>
 <button id="badminton">배드민턴</button>
 <!--이제 버튼을 누르면, sportsid가 바뀜.-->
 
-<div id="history"></div>
+<div id="sportstype"></div>
 
 </body>
 </html>
